services/security_platform/pyproject.toml
[tool.poetry]
name = "security-platform"
version = "0.1.0"
description = "Security Platform microservice – JWT-based authentication migrated from Java AuthController"
authors = []
packages = [{ include = "security_platform", from = "." }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
pydantic = "^2.7"
PyJWT = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
httpx = "^0.27"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
